requirements.txt
aiohttp
openai
tiktoken
python-dotenv